+++ sdk/runanywhere-commons/src/infrastructure/model_management/model_info_make_proto.cpp
@@ -166,10 +166,6 @@ runanywhere::v1::InferenceFramework inference_framework_to_proto(rac_inference_f
             return runanywhere::v1::INFERENCE_FRAMEWORK_MLX;
         case RAC_FRAMEWORK_COREML:
             return runanywhere::v1::INFERENCE_FRAMEWORK_COREML;
-        case RAC_FRAMEWORK_METALRT:
-            return runanywhere::v1::INFERENCE_FRAMEWORK_METALRT;
-        case RAC_FRAMEWORK_GENIE:
-            return runanywhere::v1::INFERENCE_FRAMEWORK_GENIE;
         case RAC_FRAMEWORK_SHERPA:
             return runanywhere::v1::INFERENCE_FRAMEWORK_SHERPA;
         case RAC_FRAMEWORK_QHEXRT:
@@ -199,10 +195,6 @@ rac_inference_framework_t inference_framework_from_proto(runanywhere::v1::Infere
             return RAC_FRAMEWORK_MLX;
         case runanywhere::v1::INFERENCE_FRAMEWORK_COREML:
             return RAC_FRAMEWORK_COREML;
-        case runanywhere::v1::INFERENCE_FRAMEWORK_METALRT:
-            return RAC_FRAMEWORK_METALRT;
-        case runanywhere::v1::INFERENCE_FRAMEWORK_GENIE:
-            return RAC_FRAMEWORK_GENIE;
         case runanywhere::v1::INFERENCE_FRAMEWORK_SHERPA:
             return RAC_FRAMEWORK_SHERPA;
         case runanywhere::v1::INFERENCE_FRAMEWORK_QHEXRT: